+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
+## [2.0.0] - 2026-02-04
+
+### Changed
+
+- Updated to support Xaero's Minimap 25.3.x
+- General cleanup
+
+### Added
+
+- Added sub-season support for Fabric Seasons
+
 ## [1.13.17] - 2026-01-01
 
 ### Changed